Initial work on remote DOM
The only DOM we can reach from a e10s frame script is the one of the
page - there does not seem to be a way to create a "Firebug layer" below
it (nor would that be compatible with detached mode). So rather we
introduce a proxy DOM ("remote DOM") given to the frame script, which
asynchronously sends information about operations performed on it to
chrome scope, which mirrors it with real UI.

The DOM manually manages its nodes' life times, and sends signals to the
"server" (i.e. chrome scope) when references to nodes ought to be
dropped. Currently this is done with the heuristic that each node is
created, then eventually enters the DOM (once), and then possibly leaves
it later on, and nodes are created/dropped on the server on the
first/last of these points. Thus, it's a memory leak (for as long as the
panelNode lives) if a node is created but then never inserted into the
DOM. This could be fixed by e.g. creating and destroying nodes when they
enter/leave the DOM, but I don't think it's a problem in practice.

Frame scripts *do* have the ability to send synchronous messages to
chrome, but using that for everything seems likely to be too slow.
So we reserve the possibility only for hard-to-emulate operations like
querySelector and layout accessors. In theory we could get rid of those.

Not yet functional - e.g. an HTML parser is still missing.